+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Corrected wrong tooltips on advance search, empty/non-empty buttons (`OSIDB-3502`)
 * Show comment field on CVSSv3 when the score is the same but the comment is not empty (`OSIDB-3400`)
 * Use UTC time for created date on flaw list (`OSIDB-3478`)
+* Integrate CVSS Calculator on affect's CVSS editing field (`OSIDB-3434`)
 
 ## [2024.9.2]
 ### Added